move units to folder
