+++ src/ratalien/units/iTechBuild.ts
@@ -1,0 +1,9 @@
+export interface ITechBuild{
+  deps: string[];
+  desc: string[];
+  name: string;
+  energy: number;
+  cost: number;
+  time: number;
+  mtx:Array<Array<string>>;
+}
